Cart Exception Handled

+++ PromotionEngine/Interface/ICheckout.cs
@@ -3,6 +3,6 @@
     public interface ICheckout
     {
         void Scan(string item);
-        int GetTotalPrice();
+        double GetTotalPrice();
     }
 }
